refactor(domain): extract usage cluster into qfa.domain.usage_models
The usage-tracking cluster (Operation, CallStatus, CallContext,
LLMCallRecord, DistributionStats, UsageMetrics, OperationStats,
UsageStats, TenantStats, OperationUsageStats) moves to a sibling module
under `qfa.domain`. Same pattern as the existing
`qfa.domain.sensitivity_types` split.

The cluster has two halves with one shared dependency (Operation):
persistence primitives (CallContext, LLMCallRecord, CallStatus) and
aggregations (the six stats types). Co-locating them puts every type a
`/v1/usage` response can touch — plus every type the tracking adapter
writes — in one file, making the tracking story discoverable from a
single module rather than scattered across half of `models.py`.

Doc cross-references (`{py:class}`) in
`docs/architecture/03-components.md`, `05-data-model.md`, and
`docs/adr/007-separate-api-schemas.md` are rewritten to the new path.
No backwards-compat re-exports left in `models.py` per AGENTS.md.

`models.py` drops unused imports (datetime, Decimal, StrEnum, UUID,
field_serializer) since none of the remaining models need them.
