+++ tests/adapters/test_tracking_llm.py
@@ -8,10 +8,10 @@ import pytest
 
 from qfa.adapters.tracking_llm import TrackingLLMAdapter
 from qfa.domain.errors import LLMError
-from qfa.domain.models import (
+from qfa.domain.models import LLMResponse
+from qfa.domain.usage_models import (
     CallStatus,
     LLMCallRecord,
-    LLMResponse,
     Operation,
 )
 from qfa.services.call_context import call_scope
